% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hddm.R
\name{hddm}
\alias{hddm}
\title{Fit a hierarchical four-parameter DDM to data}
\usage{
hddm(
  rt = NULL,
  xs = NULL,
  estimator = "ML",
  control = ddm_args(),
  debug = FALSE
)
}
\arguments{
\item{rt}{A vector with reaction times}

\item{xs}{A vector with choices}

\item{estimator}{A string indicating use of "ML" or "Bayes", defaults to "ML"}

\item{control}{A list of control arguments}

\item{debug}{A control argument for debugging, for internal use only}

\item{id}{A vector to identify which times and choices belongs to which level 2-unit}
}
\description{
Fit a hierarchical four-parameter DDM to data
}
