% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsddm-package.R
\docType{package}
\name{nsddm-package}
\alias{nsddm}
\alias{nsddm-package}
\title{nsddmm: Fit triadic and dyadic round-robin data}
\description{
Provides functionality for maximum likelihood and Bayesian 
  estimation for the 4-parameter drift-diffusion model.
}
\keyword{internal}
