% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm}
\alias{ddm}
\title{Fit a four-parameter DDM to the data of a single person}
\usage{
ddm(
  rt = NULL,
  xs = NULL,
  estimator = "ML",
  control = ddm_args(),
  debug = FALSE
)
}
\arguments{
\item{rt}{A vector with reaction times}

\item{xs}{A vector with choices}

\item{estimator}{A string indicating use of "ML" or "Bayes", defaults to "ML"}

\item{control}{A list of control arguments}

\item{debug}{A control argument for debugging, for internal use only}
}
\description{
Fit a four-parameter DDM to the data of a single person
}
